requirements.txt
boto3
google-auth
python-dotenv
requests
google-api-python-client
PyJWT
google-cloud-iam
google-cloud-storage